+++ changelog.md
@@ -1,3 +1,8 @@
+v1.5.0
+------
+* Dimensional Rippers can now be overcharged with additional essentia to create bigger rifts
+* Dimensional Rippers can now enlarge existing flux rifts
+
 v1.4.0
 ------
 * Added Flux Capacitor, to help you manage your flux problem